+++ Source/Core/CesiumTerrainProvider.js
@@ -176,8 +176,8 @@ define([
 
                 requestMetadata();
             })
-            .otherwise(function() {
-                deferred.reject(new RuntimeError('An error occurred while loading options.url'));
+            .otherwise(function(e) {
+                deferred.reject(e);
             });
 
         function parseMetadataSuccess(data) {